Compute initial tokens for clusters with low number of tokens

+++ pkg/cassandra/config.go
@@ -49,13 +49,12 @@ func createJsonConfig(config api.CassandraConfig, serverVersion *semver.Version,
 }
 
 func addNumTokens(template *DatacenterConfig) {
-	// Even though we default to Cassandra's stock defaults for num_tokens, we need to
-	// explicitly set it because the config builder defaults to num_tokens: 1
-	// FIXME when importing existing clusters, we should not override the user's num_tokens setting – even if it's unset (1)
+	// Note: we put int64 values because even if int values can be marshaled just fine,
+	// Unstructured.DeepCopy() would reject them since int is not a supported json type.
 	if template.ServerType == api.ServerDistributionCassandra && template.ServerVersion.Major() == 3 {
-		template.CassandraConfig.CassandraYaml.PutIfAbsent("num_tokens", 256)
+		template.CassandraConfig.CassandraYaml.PutIfAbsent("num_tokens", int64(256))
 	} else {
-		template.CassandraConfig.CassandraYaml.PutIfAbsent("num_tokens", 16)
+		template.CassandraConfig.CassandraYaml.PutIfAbsent("num_tokens", int64(16))
 	}
 }
